requirements.txt
Django==1.8.3
django-extensions==1.5.5
django-filter==0.10.0
djangorestframework==3.1.3
Markdown==2.6.2
six==1.9.0
Werkzeug==0.10.4
